Implement student list and fix filtering issue
